skaterhax

+++ 3ds-prepper.sh
@@ -39,11 +39,54 @@ read -p "Target folder: " TARGET
 if [ -d ./3ds-prepper-tmp ]; then
     rm -fr ./3ds-prepper-tmp
 fi
+
+echo "" 
+echo "Enter Hax Type"
+echo " Currently supported"
+echo " - skaterhax"
+echo "================"
+read -p "Hax: " HAXXXX
+
+HAXXXX=$(echo "$HAXXXX" | tr '[:upper:]' '[:lower:]')
+
 mkdir ./3ds-prepper-tmp
+
+case $HAXXXX in
+    "skaterhax")
+        #alt branch is better lol
+        echo " Enter the path to your skaterhax zip"
+        echo " if you do not have one, one can be obtained at https://skater.nintendohomebrew.com"
+        echo "==================================================================================="
+        read -p "path to skaterhax zip: " SKATERLOC
+        echo "********************************************"
+        echo "* Trying base setup from zip ***************"
+        echo "********************************************"
+        echo "* If your zip file is incorrect/corrupt    *"
+        echo "* I can't help you (here be dragons)       *"
+        echo "********************************************"
+        echo "...extracting skaterhax..."
+        unzip -q -d $SKATERLOC ./3ds-prepper-tmp/skater/
+        echo "...installing base luma from zip..."
+        mkdir $TARGET/3ds
+        mkdir $TARGET/boot9strap
+        mkdir $TARGET/config
+        cp -fr ./3ds-prepper-tmp/skater/3ds/* $TARGET/3ds
+        cp -fr ./3ds-prepper-tmp/skater/boot9strap/* $TARGET/boot9strap
+        cp -ft ./3ds-prepper-tmp/skater/config/* $TARGET/config
+        cp -f ./3ds-prepper-tmp/skater/*.bin $TARGET
+        cp -f ./3ds-prepper-tmp/skater/*.3dsx $TARGET
+        cp -f ./3ds-prepper-tmp/skater/*.firm $TARGET
+        exit
+    ;;
+
+esac
+
+if [ $HAXXXX -ne "skaterhax" ]; then
 echo "...Downloading Luma3DS..."
 wget -q -O ./3ds-prepper-tmp/luma_latest.zip $(curl -s https://api.github.com/repos/LumaTeam/Luma3DS/releases/latest | jq --raw-output '.assets[0] | .browser_download_url') 
 echo "...Downloading boot9strap..."
 wget -q -O ./3ds-prepper-tmp/B9S_latest.zip $(curl -s https://api.github.com/repos/SciresM/boot9strap/releases/latest | jq --raw-output '.assets[3] | .browser_download_url')
+fi
 echo "...Downloading FBI..."
 wget -q -O ./3ds-prepper-tmp/FBI.3dsx https://github.com/nh-server/FBI-NH/releases/download/2.6.1/FBI.3dsx
 wget -q -O ./3ds-prepper-tmp/FBI.cia https://github.com/nh-server/FBI-NH/releases/download/2.6.1/FBI.cia
@@ -53,30 +96,37 @@ echo "...Downloading GodMode9..."
 wget -q -O ./3ds-prepper-tmp/GM9_latest.zip $(curl -s https://api.github.com/repos/d0k3/GodMode9/releases/latest | jq --raw-output '.assets[0] | .browser_download_url')
 echo "...Downloading GM9 megascript..."
 wget -q -O ./3ds-prepper-tmp/GM9_megascript.gm9 $(curl -s https://api.github.com/repos/annson20/GM9Megascript/releases/latest | jq --raw-output '.assets[0] | .browser_download_url')
+if [ $HAXXXX -ne "skaterhax" ]; then
 if [ ! -d ./3ds-prepper-tmp/luma_extr ]; then
     mkdir ./3ds-prepper-tmp/luma_extr
 fi
 if [ ! -d ./3ds-prepper-tmp/b9s_extr ]; then
     mkdir ./3ds-prepper-tmp/b9s_extr
 fi
+fi
 if [ ! -d ./3ds-prepper-tmp/gm9_extr ]; then
     mkdir ./3ds-prepper-tmp/gm9_extr
 fi
+
+if [ $HAXXXX -ne "skaterhax" ]; then
 echo "...extracting Luma3DS..."
 unzip -q -d ./3ds-prepper-tmp/luma_extr ./3ds-prepper-tmp/luma_latest.zip
 echo "...extracting boot9strap..."
 unzip -q -d ./3ds-prepper-tmp/b9s_extr ./3ds-prepper-tmp/B9S_latest.zip
+fi
 echo "...extracting GodMode9..."
 unzip -q -d ./3ds-prepper-tmp/gm9_extr ./3ds-prepper-tmp/GM9_latest.zip
 cd ./3ds-prepper-tmp
 mkdir NEWSD
 echo "...finalizing..."
+if [ $HAXXXX -ne "skaterhax" ]; then
 cp ./luma_extr/boot.* ./NEWSD
 mkdir ./NEWSD/luma
 mkdir ./NEWSD/luma/payloads
 cp ./gm9_extr/GodMode9.firm ./NEWSD/luma/payloads
 mkdir ./NEWSD/boot9strap
 mkdir ./NEWSD/3ds
+fi
 mkdir ./NEWSD/CIAs
 mkdir ./NEWSD/gm9
 mkdir ./NEWSD/gm9/scripts
